feat(Column): Column 이동및 메서드추가
- moveWorkInColumn
기존의 Work을 대상 Column으로 이동시키는 메서드

- deleteWork
해당 Work를 삭제하는 메서드
